tests: move ECH serialize/deseriaize tests to handshake_test
These are more unit tests than integration tests, and belong alongside
where we test other handshake message types.
